+++ src/splitter_mr/embedding/__init__.py
@@ -1,4 +1,9 @@
 from .base_embedding import BaseEmbedding
-from .embeddings import AzureOpenAIEmbedding, OpenAIEmbedding
+from .embeddings import AzureOpenAIEmbedding, HuggingFaceEmbedding, OpenAIEmbedding
 
-__all__ = ["BaseEmbedding", "AzureOpenAIEmbedding", "OpenAIEmbedding"]
+__all__ = [
+    "BaseEmbedding",
+    "AzureOpenAIEmbedding",
+    "OpenAIEmbedding",
+    "HuggingFaceEmbedding",
+]