+++ src/splitter_mr/model/__init__.py
@@ -1,4 +1,15 @@
-from .base_model import BaseModel
-from .models import AzureOpenAIVisionModel, OpenAIVisionModel
+from .base_model import BaseVisionModel
+from .models import (
+    AzureOpenAIVisionModel,
+    GrokVisionModel,
+    HuggingFaceVisionModel,
+    OpenAIVisionModel,
+)
 
-__all__ = ["BaseModel", "AzureOpenAIVisionModel", "OpenAIVisionModel"]
+__all__ = [
+    "BaseVisionModel",
+    "AzureOpenAIVisionModel",
+    "OpenAIVisionModel",
+    "HuggingFaceVisionModel",
+    "GrokVisionModel",
+]